Create fseek.py
So I was looking for a way to seek to a file location from the unix command line and realized it can be easily accomplished using
python !  This may prove useful for a powerful indexing strategy.  For example word based indexing similar to model M204 on the IBM main frame.